+++ src/main/java/org/sopt/collaboration/domain/club/controller/ClubController.java
@@ -3,6 +3,7 @@ package org.sopt.collaboration.domain.club.controller;
 import lombok.RequiredArgsConstructor;
 
 import org.sopt.collaboration.domain.club.dto.ClubDto;
+import org.sopt.collaboration.domain.club.dto.ClubRecruitDto;
 import org.sopt.collaboration.domain.club.service.ClubService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,8 +17,13 @@ import java.util.List;
 public class ClubController {
 	private final ClubService clubService;
 
-	@GetMapping("/rank")
+	@GetMapping(path = "/rank")
 	public List<ClubDto> getClubsByLikes() {
 		return clubService.getClubsByLikesDesc();
+	}
+
+	@GetMapping(path = "/popular")
+	public List<ClubRecruitDto> getClubRecruitByViewCount(){
+		return clubService.getClubRecruitByViewCount();
 	}
 }
