+++ src/main/java/org/sopt/collaboration/domain/club/repository/ClubRecruitRepository.java
@@ -1,4 +1,4 @@
-package org.sopt.collaboration.domain.clubRecruit.repository;
+package org.sopt.collaboration.domain.club.repository;
 
 import org.sopt.collaboration.domain.club.entity.ClubRecruit;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ClubRecruitRepository extends JpaRepository<ClubRecruit, Long> {
-    List<ClubRecruit> findTop5ByOrderByViewCountDesc();
+	List<ClubRecruit> findTop5ByOrderByViewCountDesc();
 }
