+++ src/main/java/org/sopt/collaboration/domain/club/service/ClubService.java
@@ -1,7 +1,10 @@
 package org.sopt.collaboration.domain.club.service;
 
 import lombok.RequiredArgsConstructor;
+
 import org.sopt.collaboration.domain.club.dto.ClubDto;
+import org.sopt.collaboration.domain.club.dto.ClubRecruitDto;
+import org.sopt.collaboration.domain.club.repository.ClubRecruitRepository;
 import org.sopt.collaboration.domain.club.repository.ClubRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,12 +15,20 @@ import java.util.List;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ClubService {
-    private final ClubRepository clubRepository;
+	private final ClubRepository clubRepository;
+	private final ClubRecruitRepository clubRecruitRepository;
 
-    public List<ClubDto> getClubsByLikesDesc(){
-        return clubRepository.findTop3ByOrderByLikeCountDesc()
-                .stream()
-                .map(ClubDto::new)
-                .toList();
-    }
+	public List<ClubDto> getClubsByLikesDesc() {
+		return clubRepository.findTop3ByOrderByLikeCountDesc()
+			.stream()
+			.map(ClubDto::new)
+			.toList();
+	}
+
+	public List<ClubRecruitDto> getClubRecruitByViewCount() {
+		return clubRecruitRepository.findTop5ByOrderByViewCountDesc()
+			.stream()
+			.map(ClubRecruitDto::new)
+			.toList();
+	}
 }
